Remove `assert-news` from pre-commit checks
The `assert-news` with option `-l` is designed with an assumption that
there will only be one commit per PR, therefore it expects a newsfile
for every commit which is incorrect. A newsfile is expected per PR.
Therefore removing `assert-news` from pre-commit checks.

CI will still run `assert-news` on every PR and will fail if newsfile is
missing in the PR.

Signed-off-by: Devaraj Ranganna <devaraj.ranganna@arm.com>